+++ rlinf/utils/weight_syncer/__init__.py
@@ -1,0 +1,30 @@
+# Copyright 2026 The RLinf Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .base import WeightSyncer
+from .bucket_syncer import BucketWeightSyncer
+from .compressor import IdentityCompressor, NVCompCompressor, PatchCompressor
+from .patch_syncer import CompressedWeightPatch, PatchWeightSyncer, WeightPatch, WeightPatchTransport
+
+__all__ = [
+    "BucketWeightSyncer",
+    "CompressedWeightPatch",
+    "IdentityCompressor",
+    "NVCompCompressor",
+    "PatchCompressor",
+    "PatchWeightSyncer",
+    "WeightPatch",
+    "WeightPatchTransport",
+    "WeightSyncer",
+]
